unify shell and process sbt dump somewhat #SCL-11525
